requirements.txt
# Course requirements
python-dotenv
streamlit

# Tools

# For hello milo, blog, etc.
openai

# for the poc
jupyter

# for the video
anthropic
youtube-transcript-api

# for the blog
# Make sure you have chrome browser installed
selenium
webdriver-manager
html2text
llama-index


llama-cpp-python
nltk
langchain
chromadb
gdown
pandas
tqdm
watchdog

# Code QA
black
mypy
ruff
pytest
isort
blacken-docs
pyupgrade
detect-secrets
tomli
pre-commit
